+++ checkpoint/orbax/checkpoint/experimental/v1/_src/layout/numpy_layout.py
@@ -189,6 +189,7 @@ class NumpyLayout(CheckpointLayout):
     commit_timestamp_nsecs = int(stat_result.mtime * 1e9)
 
     return metadata_types.CheckpointMetadata[dict[str, Any]](
+        path=path,
         metadata={checkpoint_layout.PYTREE_CHECKPOINTABLE_KEY: metadata_tree},
         commit_timestamp_nsecs=commit_timestamp_nsecs,
     )